Created initial segmented collection method
This method allows us to collect chunks of the entire dataset that can eventually be combined, rather than attempting to do it all at once. However, there is still an issue of request limiting that I'm not quite sure how to deal with yet.
One option would be to stop making requests as soon as a certain number fail, and then wait some period of time.